test: move segment logging tests to SegmentLoggingTests
