+++ Tests/AWSXRayRecorderTests/SegmentTests.swift
@@ -39,78 +39,6 @@ final class SegmentTests: XCTestCase {
         }
     }
 
-    // MARK: State
-
-    func testLoggingErrors() {
-        let logHandler = TestLogHandler()
-        let logger = Logger(label: "test", factory: { _ in logHandler })
-
-        let now = Date().timeIntervalSince1970
-        let startTime = Timestamp(secondsSinceEpoch: now)!
-        let beforeTime = Timestamp(secondsSinceEpoch: now - 1)!
-
-        var numErrors: Int = 0
-
-        let segment = Segment(id: .init(), name: UUID().uuidString, context: .init(), baggage: .init(),
-                              startTime: startTime,
-                              logger: logger)
-
-        // cannot emit if still in progress
-        XCTAssertThrowsError(try segment.emit()) { error in
-            guard case SegmentError.inProgress = error else {
-                XCTFail()
-                return
-            }
-            numErrors += 1
-        }
-
-        // cannot end before started
-        XCTAssertThrowsError(try segment.end(beforeTime)) { error in
-            guard case SegmentError.backToTheFuture = error else {
-                XCTFail()
-                return
-            }
-            numErrors += 1
-        }
-
-        XCTAssertNoThrow(try segment.end(Timestamp()))
-
-        // cannot end if already end
-        XCTAssertThrowsError(try segment.end(Timestamp())) { error in
-            guard case SegmentError.alreadyEnded = error else {
-                XCTFail()
-                return
-            }
-            numErrors += 1
-        }
-        // public API does not throw but should log the error
-        segment.end()
-        numErrors += 1
-
-        XCTAssertNoThrow(try segment.emit())
-
-        // cannot emit twice
-        XCTAssertThrowsError(try segment.emit()) { error in
-            guard case SegmentError.alreadyEmitted = error else {
-                XCTFail()
-                return
-            }
-            numErrors += 1
-        }
-
-        XCTAssertEqual(numErrors, logHandler.errorMessages.count)
-
-        // log error if not emitted
-        var notEmittedSegment: Segment? = Segment(id: .init(), name: UUID().uuidString,
-                                                  context: .init(), baggage: .init(),
-                                                  logger: logger)
-        XCTAssertTrue(notEmittedSegment!.isSampled)
-        notEmittedSegment = nil
-        numErrors += 1
-
-        XCTAssertEqual(numErrors, logHandler.errorMessages.count)
-    }
-
     // MARK: Subsegments
 
     func testSubsegmentsInProgress() {
